Added comments and likes

+++ graphql/resolvers/posts.js
@@ -1,4 +1,4 @@
-const { AuthenticationError } = require('apollo-server')
+const { AuthenticationError, UserInputError } = require('apollo-server')
 
 const Post = require('../../models/Post')
 const checkAuth = require('../../util/check-auth')
@@ -47,7 +47,7 @@ module.exports = {
 			
 			try {
 				const post = await Post.findById(postId)
-				if (user.username === post.username) {
+				if (user !== null && user.username === post.username) {
 					await post.delete()
 					return 'Post deleted successfully'
 				} else {
@@ -56,6 +56,26 @@ module.exports = {
 			} catch(err) {
 				throw new Error(err)
 			}
-		}
+		},
+		async likePost(_, { postId }, context) {
+			const { username } = checkAuth(context)
+
+			const post = await Post.findById(postId)
+			if (post) {
+				if (post.likes.find(like => like.username === username)) {
+					// If post is already liked, unlike it
+					post.likes = post.likes.filter(like => like.username !== username)
+				} else {
+					// In not liked, like
+					post.likes.push({
+						username,
+						createdAt: new Date().toISOString()
+					})
+				}
+
+				await post.save()
+				return post
+			} else throw new UserInputError('Post not found')
+        }
     }
 }
